add condition on deleteFile

+++ src/modules/movie/movieController.js
@@ -157,7 +157,9 @@ module.exports = {
           delete setData[data];
         }
       });
-      deleteFile(`public/uploads/movie/${checkId[0].image}`);
+      if (req.file && checkId[0].image) {
+        deleteFile(`public/uploads/movie/${checkId[0].image}`);
+      }
       const result = await movieModel.updateMovie(setData, id);
       return helperWrapper.response(res, 200, "Success update data", result);
     } catch (error) {
@@ -181,7 +183,9 @@ module.exports = {
           null
         );
       }
-      deleteFile(`public/uploads/movie/${checkId[0].image}`);
+      if (req.file && checkId[0].image) {
+        deleteFile(`public/uploads/movie/${checkId[0].image}`);
+      }
       const result = await movieModel.deleteMovie(id);
       return helperWrapper.response(
         res,
